Returner ren JSON fra DevelopmentController (#23)
* Returner JSON-response direkte

* Bruk egne klasser

+++ src/main/kotlin/no/nav/helse/flex/arbeidssokerregister/KafkaKeyGeneratorClient.kt
@@ -28,5 +28,5 @@ data class KafkaKeyGeneratorRequest(
 )
 
 data class KafkaKeyGeneratorResponse(
-    val key: Int,
+    val key: Long,
 )
